requirements.txt
google-cloud-bigquery==3.25.0
pandas==2.2.0
pyarrow==15.0.0
db-dtypes==1.2.0
PyYAML==6.0.1
oracledb==2.2.1
